+++ bigquery-connector-common/src/main/java/com/google/cloud/bigquery/connector/common/BigQueryUtil.java
@@ -71,6 +71,7 @@ import java.util.Locale;
 import java.util.Map;
 import java.util.Optional;
 import java.util.TreeMap;
+import java.util.UUID;
 import java.util.function.Function;
 import java.util.function.IntFunction;
 import java.util.regex.Matcher;
@@ -96,8 +97,10 @@ public class BigQueryUtil {
 
   static final String READ_SESSION_EXPIRED_ERROR_MESSAGE = "session expired at";
 
-  private static final String TARGET_ALIAS = "__target";
-  private static final String SOURCE_ALIAS = "__source";
+  private static final String TARGET_ALIAS =
+      "__target_" + UUID.randomUUID().toString().replace("-", "");
+  private static final String SOURCE_ALIAS =
+      "__source_" + UUID.randomUUID().toString().replace("-", "");
 
   private static final String PROJECT_PATTERN = "\\S+";
   private static final String DATASET_PATTERN = "\\w+";
@@ -797,14 +800,14 @@ public class BigQueryUtil {
     String commaSeparatedSourceFields =
         allFields.stream()
             .map(Field::getName)
-            .map(name -> String.format("%s.`%s`", sourceAlias, name))
+            .map(name -> String.format("`%s`.`%s`", sourceAlias, name))
             .collect(Collectors.joining(","));
 
     String queryFormat =
         "DECLARE partitions_to_delete DEFAULT "
             + "(SELECT ARRAY_AGG(DISTINCT(%s) IGNORE NULLS) FROM `%s`); \n"
-            + "MERGE `%s` AS %s\n"
-            + "USING `%s` AS %s\n"
+            + "MERGE `%s` AS `%s`\n"
+            + "USING `%s` AS `%s`\n"
             + "ON FALSE\n"
             + "WHEN NOT MATCHED BY SOURCE AND (%s) AND %s IN UNNEST(partitions_to_delete) THEN DELETE\n"
             + "WHEN NOT MATCHED BY TARGET THEN\n"
@@ -841,13 +844,13 @@ public class BigQueryUtil {
             partitionField, end, partitionField, start, end, interval);
     String extractedPartitionedTarget =
         String.format(
-            "IFNULL(IF(%s.%s >= %s, 0, RANGE_BUCKET(%s.%s, GENERATE_ARRAY(%s, %s, %s))), -1)",
+            "IFNULL(IF(`%s`.`%s` >= %s, 0, RANGE_BUCKET(`%s`.`%s`, GENERATE_ARRAY(%s, %s, %s))), -1)",
             TARGET_ALIAS, partitionField, end, TARGET_ALIAS, partitionField, start, end, interval);
     // needed for tables that require the partition field to be in the where clause. It must be
     // true.
     String partitionMatchAdditionalCondition =
         String.format(
-            "%s.%s is NULL OR %s.%s >= %d",
+            "`%s`.`%s` is NULL OR `%s`.`%s` >= %d",
             TARGET_ALIAS, partitionField, TARGET_ALIAS, partitionField, Long.MIN_VALUE);
 
     return createOptimizedMergeQuery(