+++ CHANGES.md
@@ -2,7 +2,9 @@
 
 ## Next
 
-* Fix a bug where user columns named `target` or `source` cause BigQuery `MERGE` statement ambiguity on overwrite.
+* Issue #1476: Fixed BigQuery MERGE statement ambiguity on overwrite when table columns are named `target` or `source`.
+
+## 0.44.1 - 2026-03-25
 * BigQuery API has been upgraded to version 2.60.0
 * BigQuery Storage API has been upgraded to version 3.22.1
 * GAX has been upgraded to version 2.75.0
@@ -10,6 +12,7 @@
 * Guava has been upgraded to version 33.5.0-jre
 * Netty has been upgraded to version 4.2.10.Final
 * Protocol Buffers has been upgraded to version 4.34.0
+* google-api-client has been upgraded to version 2.9.0
 
 ## 0.44.0 - 2026-02-11
 * Added new connector, `spark-4.1-bigquery` aimed to be used in Spark 4.1. Like Spark 4.1, this connector requires at